+++ app/vmui/packages/vmui/src/layouts/Header/HeaderControls/HeaderControls.tsx
@@ -18,6 +18,7 @@ export interface ControlsProps {
   isMobile?: boolean;
   headerSetup?: RouterOptionsHeader;
   accountIds?: string[];
+  closeModal: () => void;
 }
 
 const HeaderControls: FC<ControlsProps & HeaderProps> = ({
@@ -45,6 +46,7 @@ const HeaderControls: FC<ControlsProps & HeaderProps> = ({
       isMobile={isMobile}
       accountIds={accountIds}
       headerSetup={headerSetup}
+      closeModal={handleCloseList}
     />
   );